% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_ved_url}
\alias{build_ved_url}
\title{Build Vermont AOE enrollment data URL}
\usage{
build_ved_url()
}
\value{
URL string for the enrollment dataset
}
\description{
Constructs the URL to download the Vermont Education Dashboard enrollment
dataset. Vermont AOE provides a single Excel file containing all years
of data.
}
\keyword{internal}
