% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{download_organizations_data}
\alias{download_organizations_data}
\title{Download organizations dataset from Vermont AOE}
\usage{
download_organizations_data()
}
\value{
Data frame with organization data
}
\description{
Download organizations dataset from Vermont AOE
}
\keyword{internal}
