% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{download_principals_data}
\alias{download_principals_data}
\title{Download principals directory from Vermont AOE}
\usage{
download_principals_data()
}
\value{
Data frame with principal data
}
\description{
Download principals directory from Vermont AOE
}
\keyword{internal}
