% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{download_superintendents_data}
\alias{download_superintendents_data}
\title{Download superintendents directory from Vermont AOE}
\usage{
download_superintendents_data()
}
\value{
Data frame with superintendent data
}
\description{
Download superintendents directory from Vermont AOE
}
\keyword{internal}
