% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_ved_enrollment}
\alias{download_ved_enrollment}
\title{Download Vermont Education Dashboard enrollment file}
\usage{
download_ved_enrollment(url)
}
\arguments{
\item{url}{URL to the enrollment dataset}
}
\value{
Data frame with enrollment data
}
\description{
Downloads and reads the VED enrollment Excel file.
}
\keyword{internal}
