% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_available_directory_years}
\alias{get_available_directory_years}
\title{Get available years of Vermont directory data}
\usage{
get_available_directory_years()
}
\value{
Integer vector of available school years (end years)
}
\description{
Returns a vector of school year ends for which Vermont school directory
data is available through the Vermont Education Dashboard.
}
\examples{
get_available_directory_years()
}
