% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr_cached}
\alias{get_raw_enr_cached}
\title{Get Vermont enrollment data for a specific year with caching}
\usage{
get_raw_enr_cached()
}
\value{
Data frame with all available enrollment data
}
\description{
Internal function that handles caching of raw data. This is used by
the main download function to avoid repeated downloads of the same
source file.
}
\keyword{internal}
