% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_year}
\alias{parse_school_year}
\title{Parse Vermont school year string}
\usage{
parse_school_year(sy)
}
\arguments{
\item{sy}{School year string}
}
\value{
Integer end year
}
\description{
Converts Vermont school year format (e.g., "2023-2024" or "SY 2023-24")
to an end year integer (2024).
}
\keyword{internal}
